[MRJ-18] realisation des operations ajouter promotion

+++ src/main/java/ma/yc/marjane/repository/AdminCentreAuthetificationRepository.java
@@ -5,6 +5,6 @@ import org.springframework.data.jpa.repository.JpaRepository;
 import org.springframework.stereotype.Repository;
 
 @Repository
-public interface AdminCentreAuthetificationRepository extends JpaRepository<AdminCentre, Long> {
+public interface AdminCentreAuthetificationRepository extends JpaRepository<AdminCentre, Integer> {
     public AdminCentre findByEmail(String email);
 }
